Add features and architecture
