+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.ArtExploreApp"
         tools:targetApi="31">
         <activity
-            android:name=".ui.view.MainActivity"
+            android:name=".util.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.ArtExploreApp">
             <intent-filter>
