+++ app/src/main/java/com/danp/artexploreapp/auth/presentation/screens/register/RegisterScreen.kt
@@ -37,7 +37,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.danp.artexploreapp.auth.presentation.viewmodels.LoginViewModel
+import com.danp.artexploreapp.auth.presentation.viewModels.LoginViewModel
 
 @Composable
 fun LoginScreen(viewModel: LoginViewModel) {
