+++ app/src/main/java/com/danp/artexploreapp/home/presentation/screens/Home.kt
@@ -1,25 +1,18 @@
-package com.danp.artexploreapp.ui.view
+package com.danp.artexploreapp.home.presentation.screens
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.TopAppBarDefaults.topAppBarColors
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import com.danp.artexploreapp.ui.theme.GreenJC
 
-@OptIn(ExperimentalMaterial3Api::class)
+
 @Composable
 fun Home(navController: NavController) {
     Box(modifier = Modifier.fillMaxSize()) {
@@ -34,23 +27,3 @@ fun Home(navController: NavController) {
         }
     }
 }
-
-//    Scaffold(
-//        topBar = {
-//            TopAppBar(
-//                colors = topAppBarColors(
-//                    containerColor = MaterialTheme.colorScheme.primaryContainer,
-//                    titleContentColor = MaterialTheme.colorScheme.primary,
-//                ),
-//                title = {
-//                    Text("Home")
-//                }
-//            )
-//        },
-//
-//    ) { innerPadding ->
-//
-//
-//    }
-
-
