+++ app/src/main/java/com/danp/artexploreapp/util/navigation/AppNavigation.kt
@@ -1,14 +1,21 @@
-package com.danp.artexploreapp.ui.view
+package com.danp.artexploreapp.util.navigation
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.navigation.NavController
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import com.danp.artexploreapp.auth.presentation.screens.login.LoginScreen
-import com.danp.artexploreapp.auth.presentation.viewmodels.LoginViewModel
-import com.danp.artexploreapp.ui.viewmodel.GalleryViewModel
+import com.danp.artexploreapp.auth.presentation.viewModels.LoginViewModel
+import com.danp.artexploreapp.home.presentation.screens.Home
+import com.danp.artexploreapp.googleMap.presentation.screens.Map
+import com.danp.artexploreapp.GalleryArt.presentation.screens.MapMuseum
+import com.danp.artexploreapp.profile.presentation.Perfil
+import com.danp.artexploreapp.qr.presentation.QrScreen
+import com.danp.artexploreapp.artRoom.presentation.screens.Room1
+import com.danp.artexploreapp.settings.presentation.screens.Settings
+import com.danp.artexploreapp.auth.presentation.screens.register.SignUp
+import com.danp.artexploreapp.artRoom.presentation.viewModels.ArtRoomViewModel
 
 @Composable
 fun AppNavigation(navController: NavHostController, modifier: Modifier = Modifier) {
@@ -24,7 +31,7 @@ fun AppNavigation(navController: NavHostController, modifier: Modifier = Modifie
         composable(route = Screens.ScreenUser.route) { Perfil(navController) }
 
         composable(route = Screens.ScreenSettings.route) { Settings(navController) }
-        composable(route = Screens.ScreenRoom1Map.route) {Room1(navController, GalleryViewModel(),) }
+        composable(route = Screens.ScreenRoom1Map.route) { Room1(navController, ArtRoomViewModel(),) }
